# bauth_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bauth",
    "description": "Headless authentication server. App routes answer errors with `ErrorBody` and a stable `code`; `/oauth/*` follows RFC 6749.",
    "contact": {
      "name": "Egor Berezovskiy"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/flows/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "operationId": "create_flow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Flow started: show the screens for `methods`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_client`, `invalid_redirect_uri`, `invalid_code_challenge`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/flows/login/{flow_id}/magic-code": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Completes the flow with the code of its magic link email, typed on the device that asked for it,\ncreating the account if the email was a sign-up. Only the newest email's code works, for 5 wrong\nattempts (10 per address a day). A wrong code, an unknown flow and no email sent all answer\n`invalid_code`.",
        "operationId": "confirm_magic_code",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "description": "Returned by `POST /flows/login`",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Exchange `code` on `/oauth/token` with the flow's `code_verifier`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_code`, `account_disabled`, `flow_expired`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/flows/login/{flow_id}/magic-link": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Emails a link and a 6-digit code for this flow: to log in if the account exists, or to create\nit (verified) when the client allows sign-up. Nothing is created before the\nlink or code is used. Always answers the same way. A new email disables the previous code.",
        "operationId": "request_magic_link",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "description": "Returned by `POST /flows/login`",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Same answer whether the account exists, is created on first use, or no email is sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicLinkResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `flow_expired`, `invalid_client`, `invalid_email`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`; after 3 emails on one flow, until the flow expires: start a new one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/magic-link/confirm": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Called by the app page the link opens. Completes the login flow the link was requested from,\ncreating the account if the email was a sign-up.",
        "operationId": "confirm_magic_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Exchange `code` on `/oauth/token` with the flow's `code_verifier`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_token`, `account_disabled`, `flow_expired`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/oauth/token": {
      "post": {
        "tags": [
          "OAuth"
        ],
        "operationId": "token",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_grant`, `unsupported_grant_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "`invalid_client`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConfirmRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "CreateFlowRequest": {
        "type": "object",
        "required": [
          "client_id",
          "redirect_uri",
          "code_challenge",
          "code_challenge_method"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "code_challenge": {
            "type": "string"
          },
          "code_challenge_method": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "state": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque value echoed back to the app with the code (CSRF protection on redirects)."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Error of every app route. Apps translate `code`, which is stable; `message` is for developers.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "invalid_code"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "FlowResponse": {
        "type": "object",
        "required": [
          "flow_id",
          "methods",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "flow_id": {
            "type": "string",
            "format": "uuid"
          },
          "methods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoginMethod"
            }
          }
        }
      },
      "LoginMethod": {
        "type": "string",
        "enum": [
          "magic_link"
        ]
      },
      "LoginResponse": {
        "oneOf": [
          {
            "type": "object",
            "description": "Exchange `code` on `/oauth/token` with the PKCE `code_verifier`.",
            "required": [
              "code",
              "status"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "completed"
                ]
              }
            }
          }
        ],
        "description": "What a login method returns once the user is authenticated."
      },
      "MagicCodeRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The 6 digits of the magic link email, without spaces.",
            "example": "042917"
          }
        }
      },
      "MagicLinkRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "MagicLinkResponse": {
        "type": "object",
        "required": [
          "status",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the flow expires now: each email keeps it alive as long as itself. Same value whether\nthe account exists or not (it only depends on the flow)."
          },
          "status": {
            "$ref": "#/components/schemas/MagicLinkStatus"
          }
        }
      },
      "MagicLinkStatus": {
        "type": "string",
        "enum": [
          "magic_link_sent"
        ]
      },
...
      "OAuthErrorBody": {
        "type": "object",
        "description": "RFC 6749 §5.2 error.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "error_description": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TokenRequest": {
        "type": "object",
        "description": "`application/x-www-form-urlencoded`, as required by RFC 6749 §4.1.3.",
        "required": [
          "grant_type",
          "client_id"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "code_verifier": {
            "type": [
              "string",
              "null"
            ]
          },
          "grant_type": {
            "type": "string"
          },
          "redirect_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "refresh_token": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "refresh_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from `/oauth/token`"
      }
    }
  }
}